+++ view/frontend/web/js/view/monri_wspay.js
@@ -4,7 +4,7 @@
  * (c) Monri Payments d.o.o.
  *
  * @author Favicode <contact@favicode.net>
- * @version 1.5.1
+ * @version 1.6.0
  */
 
 define([